Always Use Enable BeanParam Support
