+++ jaxrs2/src/test/java/feign/jaxrs2/JAXRS2ContractTest.java
@@ -13,18 +13,20 @@
  */
 package feign.jaxrs2;
 
-import static feign.assertj.FeignAssertions.assertThat;
-import org.assertj.core.data.MapEntry;
+import feign.MethodMetadata;
+import feign.jaxrs.JAXRSContract;
+import feign.jaxrs.JAXRSContractTest;
+import feign.jaxrs2.JAXRS2ContractTest.Jaxrs2BeanParamInternals.BeanParamInput;
+import feign.jaxrs2.JAXRS2ContractTest.Jaxrs2Internals.Input;
 import org.junit.Test;
 import javax.ws.rs.*;
 import javax.ws.rs.container.AsyncResponse;
 import javax.ws.rs.container.Suspended;
 import javax.ws.rs.core.Context;
 import javax.ws.rs.core.UriInfo;
-import feign.MethodMetadata;
-import feign.jaxrs.JAXRSContract;
-import feign.jaxrs.JAXRSContractTest;
-import feign.jaxrs2.JAXRS2ContractTest.Jaxrs2Internals.Input;
+import static feign.assertj.FeignAssertions.assertThat;
+import static java.util.Arrays.asList;
+import static org.assertj.core.data.MapEntry.entry;
 
 /**
  * Tests interfaces defined per {@link JAXRS2Contract} are interpreted into expected
@@ -47,11 +49,55 @@ public class JAXRS2ContractTest extends JAXRSContractTest {
   }
 
   @Test
-  public void injectBeanParam() throws Exception {
+  public void injectWithoutBeanParam() throws Exception {
     final MethodMetadata methodMetadata =
         parseAndValidateMetadata(Jaxrs2Internals.class, "beanParameters", Input.class);
     assertThat(methodMetadata.template())
         .noRequestBody();
+  }
+
+  @Test
+  public void injectBeanParam() throws Exception {
+    final MethodMetadata methodMetadata =
+        parseAndValidateMetadata(Jaxrs2BeanParamInternals.class, "beanParameters",
+            BeanParamInput.class);
+    assertThat(methodMetadata.template())
+        .noRequestBody();
+
+    assertThat(methodMetadata.template())
+        .hasHeaders(entry("X-Custom-Header", asList("{X-Custom-Header}")));
+    assertThat(methodMetadata.template())
+        .hasQueries(entry("query", asList("{query}")));
+    assertThat(methodMetadata.formParams())
+        .isNotEmpty()
+        .containsExactly("form");
+
+  }
+
+  public interface Jaxrs2BeanParamInternals {
+    @GET
+    @Path("/")
+    void inject(@Suspended AsyncResponse ar, @Context UriInfo info);
+
+    @Path("/{path}")
+    @POST
+    void beanParameters(@BeanParam BeanParamInput beanParam);
+
+    public class BeanParamInput {
+
+      @PathParam("path")
+      String path;
+
+      @QueryParam("query")
+      String query;
+
+      @FormParam("form")
+      String form;
+
+      @HeaderParam("X-Custom-Header")
+      String header;
+
+    }
   }
